refactor(controller): decouple release from CLI
