+++ mix.exs
@@ -44,8 +44,7 @@ defmodule Orchard.MixProject do
       orchard_controller: [
         applications: [
           orchard_shared: :permanent,
-          orchard_controller: :permanent,
-          orchard_cli: :load
+          orchard_controller: :permanent
         ]
       ],
       orchard_node_agent: [
